fix(test): update unstable TestFormatDateHelper
This test depends on the time when the test is executed.
There appears that there is no way of mocking the time used
in the 'elapsed' template function, and thus the test needs
to be updated once a year.

The specific test case now runs in a separate test and it
includes a comment explaining why it will start failing
after a year has passed.